WIP bit closer to working
